requirements.txt
langchain~=0.3.26
langchain_core
langchain_community
langchain_google_genai
langchain_chroma
chromadb
langgraph~=0.5.3
fastapi~=0.116.1
uvicorn
gradio~=5.37.0
pypdf
cryptography
langchain-core~=0.3.69
pydantic~=2.11.7
python-dotenv~=1.1.1
langchain-community~=0.3.27
langchain-google-genai~=2.1.8
starlette~=0.47.1
httpx~=0.28.1